Moved panel container into its own file.
